// src/features/editCurrentUserPorfile/model/types/form.ts
export interface ProfileFormInputs {
  bio: string;
  email: string;
  discord: string;
  vk: string;
  telegram: string;
  youtube: string;
}

export enum ProfileFormInputsEnum {
  ID = 'id',
  USERNAME = 'username',
  BIO = 'bio',
  EMAIL = 'email',
  DISCORD = 'discord',
  VK = 'vk',
  TELEGRAM = 'telegram',
  YOUTUBE = 'youtube',
}